refactor: MimeDomHelper
